requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.9.3
numpy==2.2.0
packaging==24.2
pillow==11.0.0
PyMuPDF==1.25.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
six==1.17.0
typing_extensions==4.12.2
wordcloud==1.9.4
